+++ pom.xml
@@ -9,7 +9,6 @@
     <version>1.0-SNAPSHOT</version>
 
 
-
     <packaging>war</packaging>
 
     <parent>
@@ -57,36 +56,14 @@
         <!-- MySQL database driver -->
         <mysql-connector-java.version>5.1.28</mysql-connector-java.version>
 
-        <!-- Database connection pool -->
-        <commons-dbcp.version>1.4</commons-dbcp.version>
-
         <!-- Liquibase -->
         <liquibase-core.version>3.2.2</liquibase-core.version>
-
-        <!-- Spring Framework dependencies -->
-        <spring-context.version>4.0.7.RELEASE</spring-context.version>
-        <spring-tx.version>4.0.7.RELEASE</spring-tx.version>
-        <spring-orm.version>4.0.7.RELEASE</spring-orm.version>
-        <spring-data-jpa.version>1.5.2.RELEASE</spring-data-jpa.version>
-
-        <!-- Hibernate dependencies -->
-        <hibernate-core.version>4.3.6.Final</hibernate-core.version>
-        <hibernate-entitymanager.version>4.3.6.Final</hibernate-entitymanager.version>
-        <hibernate-validator.version>4.3.2.Final</hibernate-validator.version>
-        <hibernate-envers.version>4.3.6.Final</hibernate-envers.version>
-
-        <!-- Logging dependencies -->
-        <slf4j-api.version>1.7.7</slf4j-api.version>
-        <jcl-over-slf4j.version>1.7.0</jcl-over-slf4j.version>
-        <slf4j-log4j12.version>1.7.7</slf4j-log4j12.version>
-        <log4j.version>1.2.17</log4j.version>
 
         <guava.version>18.0</guava.version>
 
         <!-- Testing lib dependencies -->
         <junit.version>4.11</junit.version>
         <mockito-all.version>1.9.5</mockito-all.version>
-        <spring-test.version>4.0.3.RELEASE</spring-test.version>
 
     </properties>
 
@@ -97,7 +74,6 @@
             <artifactId>commons-lang</artifactId>
             <version>${commons-lang.version}</version>
         </dependency>
-
 
         <!-- HSQLDB dependency - used as embedded database -->
         <dependency>
@@ -112,103 +88,12 @@
             <version>${mysql-connector-java.version}</version>
         </dependency>
 
-        <!-- Database connection pool -->
-        <dependency>
-            <groupId>commons-dbcp</groupId>
-            <artifactId>commons-dbcp</artifactId>
-            <version>${commons-dbcp.version}</version>
-        </dependency>
-
         <!-- Liquibase -->
         <dependency>
             <groupId>org.liquibase</groupId>
             <artifactId>liquibase-core</artifactId>
             <version>${liquibase-core.version}</version>
         </dependency>
-
-        <!-- Spring Framework dependencies -->
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-context</artifactId>
-            <version>${spring-context.version}</version>
-            <exclusions>
-                <exclusion>
-                    <groupId>commons-logging</groupId>
-                    <artifactId>commons-logging</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-orm</artifactId>
-            <version>${spring-orm.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-tx</artifactId>
-            <version>${spring-tx.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.data</groupId>
-            <artifactId>spring-data-jpa</artifactId>
-            <version>${spring-data-jpa.version}</version>
-        </dependency>
-
-
-        <!-- Hibernate dependencies -->
-        <dependency>
-            <groupId>org.hibernate</groupId>
-            <artifactId>hibernate-core</artifactId>
-            <version>${hibernate-core.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.hibernate</groupId>
-            <artifactId>hibernate-entitymanager</artifactId>
-            <version>${hibernate-entitymanager.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.hibernate</groupId>
-            <artifactId>hibernate-validator</artifactId>
-            <version>${hibernate-validator.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.hibernate</groupId>
-            <artifactId>hibernate-envers</artifactId>
-            <version>${hibernate-envers.version}</version>
-        </dependency>
-
-        <!-- Logging dependencies -->
-        <!--dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-api</artifactId>
-            <version>${slf4j-api.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>jcl-over-slf4j</artifactId>
-            <version>${jcl-over-slf4j.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-log4j12</artifactId>
-            <version>${slf4j-log4j12.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>log4j</groupId>
-            <artifactId>log4j</artifactId>
-            <version>${log4j.version}</version>
-        </dependency-->
 
         <dependency>
             <groupId>com.google.guava</groupId>
@@ -231,24 +116,18 @@
         </dependency>
 
         <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-test</artifactId>
-            <version>${spring-test.version}</version>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
 
         <dependency>
             <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-security</artifactId>
+            <artifactId>spring-boot-starter-data-jpa</artifactId>
         </dependency>
+
         <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-context-support</artifactId>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-test</artifactId>
         </dependency>
 
     </dependencies>
