--- src/main/java/Janis.java
@@ -1,6 +1,0 @@
-/**
- * Created by janis on 10/26/15.
- */
-public class Janis {
-    String hello="janis say hello";
-}
