--- src/main/java/lv/javaguru/java3/config/AppPropertiesConfig.java
@@ -1,35 +1,0 @@
-package lv.javaguru.java3.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.Resource;
-
-@Configuration
-public class AppPropertiesConfig {
-
-    @Bean
-    public static PropertySourcesPlaceholderConfigurer prodPropertiesPlaceholderConfigurer() {
-        PropertySourcesPlaceholderConfigurer p = new PropertySourcesPlaceholderConfigurer();
-        Resource[] resourceLocations = new Resource[] {
-                new ClassPathResource("database.properties")
-        };
-        p.setLocations(resourceLocations);
-        return p;
-    }
-
-/*
-    @Bean
-    @Profile("test")
-    public static PropertySourcesPlaceholderConfigurer testPropertiesPlaceholderConfigurer() {
-        PropertySourcesPlaceholderConfigurer p = new PropertySourcesPlaceholderConfigurer();
-        Resource[] resourceLocations = new Resource[] {
-                new ClassPathResource("database-test.properties")
-        };
-        p.setLocations(resourceLocations);
-        return p;
-    }
-*/
-
-}
