+++ src/main/java/lv/javaguru/java3/config/HibernateConfig.java
@@ -1,45 +1,26 @@
 package lv.javaguru.java3.config;
 
 import org.hibernate.SessionFactory;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.orm.jpa.EntityScan;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.orm.hibernate4.LocalSessionFactoryBean;
 
-import javax.sql.DataSource;
-import java.util.Properties;
+import javax.persistence.EntityManagerFactory;
 
 @Configuration
+@EntityScan(basePackages = "lv.javaguru.java3.core.domain")
 public class HibernateConfig {
 
-    @Bean
-    public Properties hibernateProperties(
-            @Value("${hibernate.dialect}") String dialect,
-            @Value("${hibernate.show_sql}") boolean showSql,
-            @Value("${hibernate.format_sql}") boolean formatSql,
-            @Value("${hibernate.hbm2ddl.auto}") String hbm2ddl) {
-
-        Properties properties = new Properties();
-        properties.put("hibernate.dialect", dialect);
-        properties.put("hibernate.show_sql", showSql);
-        properties.put("hibernate.format_sql", formatSql);
-        properties.put("hibernate.hbm2ddl.auto", hbm2ddl);
-
-        return properties;
-    }
+    @Autowired
+    private EntityManagerFactory entityManagerFactory;
 
     @Bean
-    public SessionFactory sessionFactory(DataSource dataSource,
-                                         @Value("${hibernate.packagesToScan}") String packagesToScan,
-                                         @Qualifier("hibernateProperties") Properties properties) throws Exception {
-
-        LocalSessionFactoryBean sessionFactoryBean = new LocalSessionFactoryBean();
-        sessionFactoryBean.setDataSource(dataSource);
-        sessionFactoryBean.setPackagesToScan(packagesToScan);
-        sessionFactoryBean.setHibernateProperties(properties);
-        sessionFactoryBean.afterPropertiesSet();
-        return sessionFactoryBean.getObject();
+    public SessionFactory getSessionFactory() {
+        if (entityManagerFactory.unwrap(SessionFactory.class) == null) {
+            throw new NullPointerException("factory is not a hibernate factory");
+        }
+        return entityManagerFactory.unwrap(SessionFactory.class);
     }
 
 }
