--- src/main/java/lv/javaguru/java3/config/liquibase/DatabaseMigrationException.java
@@ -1,8 +1,0 @@
-package lv.javaguru.java3.config.liquibase;
-
-public class DatabaseMigrationException extends RuntimeException {
-
-	public DatabaseMigrationException(String message, Throwable cause) {
-		super(message, cause);
-	}
-}
