+++ src/main/resources/application.properties
@@ -1,0 +1,18 @@
+# DATASOURCE
+spring.datasource.url=jdbc:hsqldb:mem:javaee_idea_keeper
+spring.datasource.username=sa
+spring.datasource.password=
+spring.datasource.initial-size=10
+
+# JPA (JpaBaseConfiguration, HibernateJpaAutoConfiguration)
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.HSQLDialect
+spring.jpa.database-platform=org.hibernate.dialect.HSQLDialect
+spring.jpa.properties.hibernate.current_session_context_class=org.springframework.orm.hibernate4.SpringSessionContext
+
+# LIQUIBASE (LiquibaseProperties)
+liquibase.change-log=classpath:/liquibase/db.changelog-master.xml
+liquibase.check-change-log-location=true
+liquibase.drop-first=false
+liquibase.enabled=true
