--- src/main/resources/database.properties
@@ -1,22 +1,0 @@
-database.driverClass=org.hsqldb.jdbc.JDBCDriver
-database.jdbcUrl=jdbc:hsqldb:mem:javaee_idea_keeper
-database.userName=Anna
-database.password=anna2302
-
-liquibase.update=true
-
-hibernate.dialect=org.hibernate.dialect.HSQLDialect
-hibernate.show_sql=true
-hibernate.format_sql=true
-hibernate.hbm2ddl.auto=validate
-hibernate.packagesToScan=lv.javaguru.java3.core.domain
-
-
-#spring.datasource.url=jdbc:hsqldb:mem:javaee_idea_keeper
-#spring.datasource.username=Anna
-#spring.datasource.password=anna2302
-#spring.datasource.driver-class-name=com.mysql.jdbc.Driver
-#spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
-#spring.jpa.show-sql = true
-#spring.jpa.hibernate.ddl-auto = update
-#spring.jpa.hibernate.naming-strategy = org.hibernate.cfg.ImprovedNamingStrategy
