+++ src/test/java/lv/javaguru/java3/core/database/users/UserDAOImplTest.java
@@ -3,6 +3,8 @@ package lv.javaguru.java3.core.database.users;
 /**
  * Created by Anna on 27.10.2015.
  */
+
+import lv.javaguru.java3.core.database.DatabaseHibernateTest;
 import lv.javaguru.java3.core.domain.user.User;
 import org.junit.Test;
 import org.springframework.transaction.annotation.Transactional;
